【部门ID到部门名称的转换】实现岗位表的 `deptId`与部门表的 `deptName`的关联显示和操作【4.0】

+++ sql/mysql/aioveu_boot-20250813-5_add-aioveu-salary.sql
@@ -16,7 +16,7 @@ DROP TABLE IF EXISTS `aioveu_salary`;
 -- 创建规范化的工资表
 CREATE TABLE `aioveu_salary` (
                                  `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT COMMENT '工资记录ID',
-                                 `employee_id` INT UNSIGNED NOT NULL COMMENT '员工ID',
+                                 `employee_id` BIGINT UNSIGNED NOT NULL COMMENT '员工ID',
                                  `salary_period` DATE NOT NULL COMMENT '工资周期（月份）', -- 使用DATE类型存储工资周期（每月1日） 添加唯一约束确保每月每人只有一条记录
                                  `base_salary` DECIMAL(12, 2) NOT NULL DEFAULT 0.00 COMMENT '基本工资',
                                  `performance_bonus` DECIMAL(12, 2) DEFAULT 0.00 COMMENT '绩效奖金',
